+++ README.md
@@ -3,7 +3,7 @@
 Music fan may have a huge amount of physical CDs, LPs, Cassettes. BandurriaJ will allow them to manage their full catalogue,
  providing tagging features and queries possibilities.
  
- BandurriaJ is a full java desktop application based on JavaFx and Spring boot.
+ BandurriaJ is a full java desktop BandurriaJ based on JavaFx and Spring boot.
  
 
 If you want to participate, please do not hesitate to contact us!
