+++ doc/design/Design.md
@@ -5,7 +5,7 @@ The core functionality of BandurriaJ is to create Catalogues of Discs, CDs and C
 fan can have a centralized place to store it and to retrieve information about it. 
 
 For instance, imagine you have a huge collection of CDs and a friend wants to know if you can borrow him
-some Jazz cd from the 70s. You can check in JBandurria and see that you have a perfect CDs with 
+some Jazz cd from the 70s. You can check in BandurriaJ and see that you have a perfect CDs with 
 the label 1002. So you can now check in your collection for the CD #1002 and lend it!.
 
 ---
