--- src/main/java/application/Controller.java
@@ -1,4 +1,0 @@
-package application;
-
-public class Controller {
-}
