+++ src/main/java/com/axokoi/BandurriaJ/CatalogueController.java
@@ -1,0 +1,4 @@
+package com.axokoi.BandurriaJ;
+
+public class CatalogueController {
+}
