+++ src/main/java/com/axokoi/BandurriaJ/cd/CdReader.java
@@ -1,4 +1,4 @@
-package application.cd;
+package com.axokoi.BandurriaJ.cd;
 
 import java.io.IOException;
 import java.io.InputStream;
