+++ src/main/java/com/axokoi/BandurriaJ/model/Band.java
@@ -1,8 +1,10 @@
-package application.model;
+package com.axokoi.BandurriaJ.model;
 
 import java.util.List;
 
+import javax.persistence.CascadeType;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
@@ -12,13 +14,19 @@ import lombok.Data;
 
 @Data
 @Entity
-public class Catalogue {
+public class Band {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private Long id;
+
 	private String name;
-	@OneToMany(targetEntity = Disc.class)
+
+	@OneToMany(targetEntity = Artist.class, cascade = CascadeType.ALL, fetch = FetchType.EAGER)
+	private List<Artist> artists;
+
+	@OneToMany(targetEntity = Disc.class, cascade = CascadeType.ALL)
 	private List<Disc> discs;
 
+	private String comment;
 
 }
