+++ src/main/java/com/axokoi/BandurriaJ/model/Catalogue.java
@@ -1,31 +1,24 @@
-package application.model;
+package com.axokoi.BandurriaJ.model;
 
 import java.util.List;
 
-import javax.persistence.CascadeType;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 
 import lombok.Data;
 
 @Data
 @Entity
-public class Disc {
+public class Catalogue {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private Long id;
 	private String name;
+	@OneToMany(targetEntity = Disc.class)
+	private List<Disc> discs;
 
-	@ManyToOne(targetEntity = Band.class, cascade = CascadeType.REFRESH)
-	private Band band;
-
-	@OneToMany(targetEntity = Track.class, cascade = CascadeType.ALL)
-	private List<Track> tracks;
-
-	private String comment;
 
 }
