+++ src/main/java/com/axokoi/BandurriaJ/model/DiscService.java
@@ -1,4 +1,4 @@
-package application.model;
+package com.axokoi.BandurriaJ.model;
 
 import java.util.Optional;
