+++ src/main/java/com/axokoi/BandurriaJ/views/CatalogueView.java
@@ -1,0 +1,4 @@
+package com.axokoi.BandurriaJ.views;
+
+public class CatalogueView {
+}
