+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.main.web-application-type=none
+spring.main.web-BandurriaJ-type=none
